add licenseOriginUrl and refactor
