+++ src/components/ModalWindow/ContentTemplates/LicenseInfo.js
@@ -1,5 +1,18 @@
 import './LicenseInfo.css';
 
-export default function LicenseInfoModalWindowContent({ licenseDetails = '' }) {
-  return <section className="license-info">{licenseDetails}</section>;
+export default function LicenseInfoModalWindowContent({ credentials: { license, licenseOriginUrl } }) {
+  return (
+    <section className="license-info">
+      <p>
+        Author's license file: <br />
+        <a href={licenseOriginUrl} target="_blank" rel="noreferrer">
+          {licenseOriginUrl}
+        </a>
+      </p>
+      <p>
+        Overview: <br />
+        {license?.body}
+      </p>
+    </section>
+  );
 }
